WIP: Add retreat phase determination
Unlike main phase and build phase, the retreat phase depends on the
orders of the previous phase in the same season to adjudicate. To
achieve this, the `judge::Outcome` struct is being used as the basis
for resolving a retreat phase, using a world state to figure out where
all the units are and using the previous phase's orders to figure out
which retreat destinations are valid.