Add convoy policy
Different editions handle convoy intent differently.
Previously, the crate implemented a single behavior; now it is specified by the rulebook.

This change was unusally invasive, as it required adding a new trait that the default calculations rely on.

It is still possible to provide a custom adjudicator, but reusing the existing calculators will now require implementing this new trait.